requirements.txt
selenium==4.5.0
python-dotenv==0.20.0
openpyxl==3.0.10
undetected_chromedriver==3.1.6